changed the timestamps to floats instead of ints

+++ sniffer.py
@@ -7,8 +7,8 @@ def monitor(packet):
   #packetlist.append(packet)
   string = json.dumps(str(packet.load))
   timestamp = str(string[3:-2])
-  nanoseconds = int(timestamp*1000000000)
-  difference = int((time.time())*1000000000) - nanoseconds
+  nanoseconds = float(timestamp*1000000000)
+  difference = float((time.time())*1000000000) - nanoseconds
   os.system("curl -i -XPOST 'http://localhost:8086/write?db=networkDatabase' --data-binary 'icmp,type=timestamps difference="+str(difference)+" "+str(nanoseconds)+"'")
   #print(string)
   #print(packet.show())
